+++ KalikoCMS.Admin/Admin/Content/PropertyType/CollectionPropertyEditor.js
@@ -6,14 +6,14 @@
                 var content = '<a href="#" onclick="top.propertyEditor.collection.editField(this);return false;" class="pull-right"><i class="icon icon-edit"></i>edit</a><i class="icon icon-sort"></i> ' + exerpt;
                 if (valueField == null && val != null) {
                     var newField = $('<li class="btn btn-default collection-item"></li>');
-                    newField.attr('data-value', escape(val)).html(content);
+                    newField.attr('data-value', encodeURIComponent(val)).html(content);
                     list.append(newField);
                 }
                 else if (val == null) {
                     $(valueField).remove();
                 }
                 else {
-                    $(valueField).attr('data-value', escape(val)).html(content);
+                    $(valueField).attr('data-value', encodeURIComponent(val)).html(content);
                 }
 
                 list.trigger('sortupdate');
@@ -23,7 +23,7 @@
 
             var value = null;
             if (valueField != null) {
-                value = unescape($(valueField).attr('data-value'));
+                value = decodeURIComponent($(valueField).attr('data-value'));
             }
             top.propertyEditor.dialogs.openEditCollectionPropertyDialog(className, value);
         },