+++ KalikoCMS.Engine/Data/Maps/PageTypeMap.cs
@@ -31,7 +31,7 @@ namespace KalikoCMS.Data.Maps {
             HasProperty(x => x.Name).ToColumn("Name").IsNullable().WithVariableLength(50);
             HasProperty(x => x.DefaultChildSortDirection).ToColumn("DefaultChildSortDirection").IsNotNullable();
             HasProperty(x => x.DefaultChildSortOrder).ToColumn("DefaultChildSortOrder").IsNotNullable();
-            HasProperty(x => x.DisplayName).ToColumn("DisplayName").IsNullable().HasColumnType("nvarchar").HasLength(50);
+            HasProperty(x => x.DisplayName).ToColumn("DisplayName").IsNullable().IsUnicode().HasLength(50);
             HasProperty(x => x.PageTypeDescription).ToColumn("PageTypeDescription").IsNullable().IsUnicode().WithVariableLength(255);
             HasProperty(x => x.PageTemplate).ToColumn("PageTemplate").IsNullable().WithVariableLength(100);
             HasProperty(x => x.ShowPublishDates).ToColumn("ShowPublishDates").IsNullable().HasDefaultValue();